arch: extract cleanup rules into pure computeCleanup() module
- Create src/background/cleanup.ts with pure computeCleanup() function
- Move getDuplicateTabs() from index.ts to cleanup.ts (re-exported)
- applyCleanupRules() now: collect Chrome data → pure compute → apply results
- Business logic is now deeply testable without Chrome mocks
- No behavior changes — all 112 existing background tests pass

Closes #9
